refactor storage lib
