+++ internal/core/src/storage/CMakeLists.txt
@@ -25,7 +25,7 @@ if (DEFINED AZURE_BUILD_DIR)
     add_definitions(-DAZURE_BUILD_DIR)
     include_directories(azure-blob-storage)
     include_directories("${AZURE_BUILD_DIR}/vcpkg_installed/${VCPKG_TARGET_TRIPLET}/include")
-    set(SOURCE_FILES ${SOURCE_FILES} azure/AzureChunkManager.cpp)
+    set(SOURCE_FILES ${SOURCE_FILES} azure-blob-storage/AzureChunkManager.cpp)
     add_library(azure_blob_chunk_manager SHARED IMPORTED)
     set_target_properties(azure_blob_chunk_manager
         PROPERTIES
@@ -43,4 +43,12 @@ if(USE_OPENDAL)
     set(SOURCE_FILES ${SOURCE_FILES} opendal/OpenDALChunkManager.cpp)
 endif()
 
+add_subdirectory(local)
+set(SOURCE_FILES ${SOURCE_FILES} local/LocalChunkManager.cpp)
+add_subdirectory(minio)
+set(SOURCE_FILES ${SOURCE_FILES} minio/MinioChunkManager.cpp)
+add_subdirectory(aliyun)
+set(SOURCE_FILES ${SOURCE_FILES} aliyun/AliyunCredentialsProvider.cpp aliyun/AliyunSTSClient.cpp)
+add_subdirectory(tencent-cloud)
+set(SOURCE_FILES ${SOURCE_FILES} tencent-cloud/TencentCloudCredentialsProvider.cpp tencent-cloud/TencentCloudSTSClient.cpp)
 add_library(milvus_storage OBJECT ${SOURCE_FILES})
