+++ internal/core/src/storage/ChunkManager.cpp
@@ -28,11 +28,11 @@
 #include <aws/s3/model/ListObjectsRequest.h>
 #include <aws/s3/model/PutObjectRequest.h>
 
-#include "storage/MinioChunkManager.h"
-#include "storage/AliyunSTSClient.h"
-#include "storage/TencentCloudSTSClient.h"
-#include "storage/AliyunCredentialsProvider.h"
-#include "storage/TencentCloudCredentialsProvider.h"
+#include "storage/minio/MinioChunkManager.h"
+#include "storage/aliyun/AliyunSTSClient.h"
+#include "storage/tencent-cloud/TencentCloudSTSClient.h"
+#include "storage/aliyun/AliyunCredentialsProvider.h"
+#include "storage/tencent-cloud/TencentCloudCredentialsProvider.h"
 #include "common/Consts.h"
 #include "common/EasyAssert.h"
 #include "log/Log.h"
