+++ internal/core/src/storage/DiskFileManagerImpl.cpp
@@ -40,7 +40,7 @@
 #include "storage/DiskFileManagerImpl.h"
 #include "storage/FileManager.h"
 #include "storage/IndexData.h"
-#include "storage/LocalChunkManagerSingleton.h"
+#include "storage/local/LocalChunkManagerSingleton.h"
 #include "storage/ThreadPools.h"
 #include "storage/Util.h"
