+++ internal/core/src/storage/aliyun/AliyunCredentialsProvider.h
