+++ internal/core/src/storage/aliyun/AliyunSTSClient.cpp
