+++ internal/core/src/storage/aliyun/AliyunSTSClient.h
