+++ internal/core/src/storage/azure-blob-storage/AzureChunkManager.cpp
@@ -21,7 +21,7 @@
 #include "common/EasyAssert.h"
 #include "log/Log.h"
 #include "monitor/prometheus_client.h"
-#include "storage/azure/AzureChunkManager.h"
+#include "AzureChunkManager.h"
 
 namespace milvus {
 namespace storage {
