+++ internal/core/src/storage/tencent-cloud/TencentCloudCredentialsProvider.cpp
