+++ internal/core/src/storage/tencent-cloud/TencentCloudSTSClient.cpp
