+++ internal/core/src/storage/tencent-cloud/TencentCloudSTSClient.h
