+++ internal/core/unittest/test_azure_chunk_manager.cpp
@@ -14,7 +14,7 @@
 #include <vector>
 
 #include "common/EasyAssert.h"
-#include "storage/azure/AzureChunkManager.h"
+#include "storage/azure-blob-storage/AzureChunkManager.h"
 #include "storage/Util.h"
 
 using namespace std;
