+++ internal/core/unittest/test_sealed.cpp
@@ -18,9 +18,9 @@
 #include "knowhere/version.h"
 #include "segcore/SegmentSealedImpl.h"
 #include "storage/MmapManager.h"
-#include "storage/MinioChunkManager.h"
+#include "storage/minio/MinioChunkManager.h"
 #include "storage/RemoteChunkManagerSingleton.h"
-#include "storage/LocalChunkManagerSingleton.h"
+#include "storage/local/LocalChunkManagerSingleton.h"
 #include "storage/Util.h"
 #include "test_utils/DataGen.h"
 #include "test_utils/indexbuilder_test_utils.h"
